+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added: Multiple memo support to `SendScene2`
 - added: Add os version to the "OS" line of logs
 - changed: "MAX" button in `FlipInput` restyled to tertiary and moved next to the text input field
+- changed: Moved spending limits settings to password-locked account settings section.
 - fixed: Primary button transforming on the animation on `GettingStartedScene`
 - fixed: `NotificationCenterScene` could crash if there was an undismissed new token notification and the associated wallet no longer existed
 - removed: Disable Fantom transaction list